Moved auxiliary functions.
